+++ src/java/model/User.java
@@ -5,25 +5,25 @@
 package model;
 
 public class User {
-    private long id;
+    private int id;
     private String name;
     private String email;
     private String birthDate;
 
     public User() {}
     
-    public User(long id, String name, String email, String birthDate) {
+    public User(int id, String name, String email, String birthDate) {
         this.id = id;
         this.name = name;
         this.email = email;
         this.birthDate = birthDate;
     }
 
-    public long getId() {
+    public int getId() {
         return id;
     }
 
-    public void setId(long id) {
+    public void setId(int id) {
         this.id = id;
     }
